Create removeUser() method inside User.
The method is not static. So if a user wants to be removed, the method needs to be called using a current existing instance. That's why I created a dummyUser  to be able to remove a user no matter what; for now.
Also, to remove a user the id is used.