Create updateUser() method
The updateUser() method needs to be called on the object that wants to be updated. The object has to already exist in the running code. When the method is called, it will take all the variables related to this object and update the corresponding row in the database. Now, if the row doesn't exist, nothing bad should happen, the method would return false since its of type boolean.